refactor: move fragment to others package
To make more organizer and better maintainability
