+++ feature/favorite/src/main/res/layout/fragment_favorite.xml
@@ -6,7 +6,7 @@
   android:layout_width="match_parent"
   android:layout_height="match_parent"
   android:orientation="vertical"
-  tools:context=".ui.FavoriteFragment">
+  tools:context=".ui.fragment.FavoriteFragment">
 
   <!-- Toolbar -->
   <com.google.android.material.appbar.AppBarLayout
